Rename view contracts
