+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/informationcard/AnalyticsCardInformationViewState.kt
@@ -1,0 +1,9 @@
+package com.woocommerce.android.ui.analytics.informationcard
+
+sealed class AnalyticsCardInformationViewState {
+    object HiddenViewState : AnalyticsCardInformationViewState()
+    data class DataViewState(val title: String,
+                             val totalValues: SectionViewState,
+                             val netValues: SectionViewState)
+        : AnalyticsCardInformationViewState()
+}
