+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/informationcard/SectionViewState.kt
@@ -1,0 +1,12 @@
+package com.woocommerce.android.ui.analytics.informationcard
+
+sealed class SectionViewState {
+    object SectionHiddenViewState : SectionViewState()
+    data class SectionDataViewState(val title: String, val value: String, val delta: Int) : SectionViewState() {
+        fun getSign(): String = when {
+            delta > 0 -> "+"
+            delta == 0 -> ""
+            else -> "-"
+        }
+    }
+}
